% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{as_dnabin}
\alias{as_dnabin}
\title{Coerce a genoaligner_msa alignment to ape::DNAbin}
\usage{
as_dnabin(x)
}
\arguments{
\item{x}{A \code{genoaligner_msa} object.}
}
\value{
An \code{ape::DNAbin} matrix.
}
\description{
Requires the \pkg{ape} package (in \code{Suggests}; used here only).
}
\examples{
if (requireNamespace("ape", quietly = TRUE)) {
    as_dnabin(msa_align(c("ACGTACGT", "ACGTTCGT")))
}
}
